pyproject.toml
[project]
name = "wip-cli"
version = "0.1.0"
description = "Where did I leave off? A morning briefing for developers."
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "gitpython>=3.1.0",
    "rich>=13.0.0",
    "tomli>=1.0.0; python_version < '3.11'",
]

[project.scripts]
wip = "wip.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/wip"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
